rename viewing_startups to view
